Fix task 0, change type of 2 variable

+++ 0x15-file_io/0-read_textfile.c
@@ -14,7 +14,7 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	ssize_t fd;
-	size_t numberWrite, numberRead;
+	int numberWrite, numberRead;
 	char *buf;
 
 	if (filename == NULL)
